cabal/haskell-overrides: fix github build
dhall has switched to using strict text, and some bounds must be
relaxed.

Also switch to using callCabal2nix, as it's easier to use and it doesn't
fall out of sync with the weird cabal mechanism that nixpkgs uses (which
might cause unfound modules due to major changes upstream, etc).

Signed-off-by: Austin Seipp <aseipp@pobox.com>